Remove more unused code.
